backend/requirements.txt
Django==4.2.5
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary
djangorestframework-simplejwt==5.2.2
boto3==1.26.94
pillow==9.4.0
django-cors-headers==3.14.0